agents/eliza-daemon/requirements.txt
# Enhanced Eliza Daemon Requirements

# Core Python packages
asyncio-mqtt==0.11.1
aiohttp==3.8.5
aiofiles==23.1.0

# AI and Machine Learning
langchain==0.0.354
openai==0.28.1
tiktoken==0.5.1

# Blockchain Integration
web3==6.11.3
eth-account==0.9.0
eth-utils==2.2.0
py-solc-x==1.12.0

# Database and Memory
supabase==1.0.4
asyncpg==0.28.4
psycopg2-binary==2.9.7

# Social Media APIs
tweepy==4.14.0
discord-webhook==1.3.0

# Scheduling and Background Tasks
APScheduler==3.10.4
celery==5.3.4
redis==5.0.1

# HTTP Requests and APIs
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.3
numpy==1.24.4
python-dateutil==2.8.2

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1

# Cryptography and Security
cryptography==41.0.7
pynacl==1.5.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# System monitoring
psutil==5.9.6
